requirements.txt

flask
cryptography
pyotp
